Move module definition
