+++ modules/layers/src/scatterplot-layer/scatterplot-layer-fragment.glsl.ts
@@ -18,14 +18,11 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import scatterplotUniforms from './scatterplot-layer-uniforms.glsl';
 export default `\
 #version 300 es
 #define SHADER_NAME scatterplot-layer-fragment-shader
 
 precision highp float;
-
-${scatterplotUniforms}
 
 in vec4 vFillColor;
 in vec4 vLineColor;
