+++ modules/layers/src/scatterplot-layer/scatterplot-layer-vertex.glsl.ts
@@ -18,7 +18,6 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import scatterplotUniforms from './scatterplot-layer-uniforms.glsl';
 export default `\
 #version 300 es
 #define SHADER_NAME scatterplot-layer-vertex-shader
@@ -33,7 +32,6 @@ in vec4 instanceFillColors;
 in vec4 instanceLineColors;
 in vec3 instancePickingColors;
 
-${scatterplotUniforms}
 uniform float opacity;
 
 out vec4 vFillColor;
